requirements.txt
GitPython
pyrogram
tgcrypto
requests
fastapi
uvicorn
PyGithub
